Removing common-ui
